Used other approach
trying back the template method and its working wonders. The program automatically click on skip intro, skip recap and next episode. Noticed that when next episode button comes out it stays on the edge of the button and doesnt click, therefore added +5 to x and y axis of the coordinate. All works perfectly now